# docs/openapi/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AskRequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "space_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Id"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationResponse"
            },
            "title": "Citations",
            "type": "array"
          },
          "rag_run_id": {
            "title": "Rag Run Id",
            "type": "string"
          },
          "real_acceptance": {
            "title": "Real Acceptance",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "rag_run_id",
          "status",
          "answer",
          "citations",
          "warnings",
          "verified",
          "real_acceptance"
        ],
        "title": "AskResponse",
        "type": "object"
      },
...
      "CitationResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "locator": {
            "additionalProperties": true,
            "title": "Locator",
            "type": "object"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "source_url",
          "locator"
        ],
        "title": "CitationResponse",
        "type": "object"
      },
...
      "CreateSpaceRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateSpaceRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "retryable"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvidenceSourceResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "locator": {
            "additionalProperties": true,
            "title": "Locator",
            "type": "object"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "text",
          "locator"
        ],
        "title": "EvidenceSourceResponse",
        "type": "object"
      },
      "FeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 2000,
            "title": "Comment",
            "type": "string"
          },
          "rating": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "type": "integer"
          },
          "reason_code": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Reason Code",
            "type": "string"
          }
        },
        "required": [
          "rating",
          "reason_code"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
      "FeedbackResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "index_generation_id": {
            "title": "Index Generation Id",
            "type": "string"
          },
          "model_revision": {
            "title": "Model Revision",
            "type": "string"
          },
          "prompt_revision": {
            "title": "Prompt Revision",
            "type": "string"
          },
          "rag_run_id": {
            "title": "Rag Run Id",
            "type": "string"
          },
          "retrieval_revision": {
            "title": "Retrieval Revision",
            "type": "string"
          }
        },
        "required": [
          "rag_run_id",
          "accepted",
          "index_generation_id",
          "retrieval_revision",
          "prompt_revision",
          "model_revision"
        ],
        "title": "FeedbackResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeDocumentResponse": {
        "additionalProperties": false,
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "parser_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Revision"
          },
          "processing_state": {
            "title": "Processing State",
            "type": "string"
          },
          "publication_state": {
            "title": "Publication State",
            "type": "string"
          },
          "space_id": {
            "title": "Space Id",
            "type": "string"
          },
          "version_id": {
            "title": "Version Id",
            "type": "string"
          },
          "version_no": {
            "title": "Version No",
            "type": "integer"
          }
        },
        "required": [
          "document_id",
          "space_id",
          "filename",
          "version_id",
          "version_no",
          "processing_state",
          "publication_state",
          "parser_revision",
          "chunk_count",
          "job_id"
        ],
        "title": "KnowledgeDocumentResponse",
        "type": "object"
      },
...
      "SearchRequest": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "anyOf": [
              {
                "maximum": 50.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "query": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Id"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
      "SearchResponse": {
        "additionalProperties": false,
        "properties": {
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/SearchHitResponse"
            },
            "title": "Hits",
            "type": "array"
          },
          "observed_security_watermark": {
            "title": "Observed Security Watermark",
            "type": "integer"
          },
          "real_acceptance": {
            "title": "Real Acceptance",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "observed_security_watermark",
          "hits",
          "real_acceptance",
          "degraded",
          "warnings"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "SpaceResponse": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "status"
        ],
        "title": "SpaceResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Enterprise RAG Knowledge Base",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ask": {
      "post": {
        "operationId": "ask_api_v1_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Ask",
        "tags": [
          "trusted-qa"
        ]
      }
    },
    "/api/v1/ask:stream": {
      "post": {
        "operationId": "ask_stream_api_v1_ask_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Ask Stream",
        "tags": [
          "trusted-qa"
        ]
      }
    },
...
    "/api/v1/rag-runs/{rag_run_id}/feedback": {
      "post": {
        "operationId": "feedback_api_v1_rag_runs__rag_run_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "rag_run_id",
            "required": true,
            "schema": {
              "title": "Rag Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Feedback",
        "tags": [
          "trusted-qa"
        ]
      }
    },
    "/api/v1/rag-runs/{run_token}/evidence/{evidence_token}/source": {
      "get": {
        "operationId": "evidence_source_api_v1_rag_runs__run_token__evidence__evidence_token__source_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_token",
            "required": true,
            "schema": {
              "title": "Run Token",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "evidence_token",
            "required": true,
            "schema": {
              "title": "Evidence Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceSourceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Evidence Source",
        "tags": [
          "trusted-qa"
        ]
      }
    },
...
    "/api/v1/search": {
      "post": {
        "operationId": "search_api_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Search",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/api/v1/spaces": {
      "get": {
        "operationId": "spaces_api_v1_spaces_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SpaceResponse"
                  },
                  "title": "Response Spaces Api V1 Spaces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Spaces",
        "tags": [
          "spaces"
        ]
      },
      "post": {
        "operationId": "create_space_api_v1_spaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Space",
        "tags": [
          "spaces"
        ]
      }
    },
    "/api/v1/spaces/{space_id}/documents": {
      "get": {
        "operationId": "documents_api_v1_spaces__space_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeDocumentResponse"
                  },
                  "title": "Response Documents Api V1 Spaces  Space Id  Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Documents",
        "tags": [
          "spaces",
          "documents"
        ]
      }
    },
...
  }
}